refactor home/models into separate package
