+++ src/home/models/__init__.py
@@ -1,0 +1,7 @@
+# flake8: noqa
+from .banner import Banner
+from .contact_page import ContactPage
+from .form_field import FormField
+from .form_page import FormPage
+from .home_page import HomePage
+from .web_page import WebPage
