+++ src/home/models/form_field.py
@@ -1,0 +1,6 @@
+from wagtail.contrib.forms.models import AbstractFormField
+from modelcluster.fields import ParentalKey
+
+
+class FormField(AbstractFormField):
+    page = ParentalKey('FormPage', related_name='form_fields')
